+++ dhis-2/dhis-services/dhis-service-tracker/src/main/java/org/hisp/dhis/tracker/export/OperationParamUtils.java
@@ -87,7 +87,7 @@ public class OperationParamUtils {
       String filterItem, Map<String, TrackedEntityAttribute> attributes)
       throws BadRequestException {
     if (StringUtils.isEmpty(filterItem)) {
-      return List.of();
+      return new ArrayList<>();
     }
 
     List<String> uidOperatorValues = filterList(filterItem);
